refactor(mocks): extract findings module from mockApi (#991 M3)
Move the incidental-findings state/helpers (findings, findingDTO,
matchFindingBucket, nextFiledIssueIid, mockIssueDraft) and its six methods
(listFindings, findingIssueDraft, fileFinding, dismissFinding, getIssueDraft,
fileIssue) verbatim into web/src/mocks/mockApi/findings.ts as findingsApi.

Cross-module reads per PRD fact 3: mockIssueDraft reads repos (import from
./forge); fileIssue/getIssueDraft read reviews (import from ./judge). After the
move noUnusedLocals dropped reviews from index.ts's ./judge import (findings was
its only reader) while repos STAYS imported in index.ts (still read by
runs/schedules/workers/admin methods).
